[MOD] INCLUSION DE SENSOR GRADOS Y MEJORAS EN LA RECEPCION.
* Si un sensor no tiene configuracion, no se lanza el TCPClient.
* Sensor value item, ahora es Float, para soportar los grados que son valores decimales.

+++ Global/Global.go
@@ -9,7 +9,7 @@ type Configuration struct {
 }
 
 type ItemInfo struct {
-	Value int	`json:"Value"`
+	Value float64	`json:"Value"`
 	Date  string  `json:"TimeReceived"`
 }
 type Store struct {
@@ -80,7 +80,7 @@ func (q *QueueItem) GetFirst() (n *ItemInfo){
 	return n
 }
 
-func (q *QueueItem) AllEqual(n int) bool {
+func (q *QueueItem) AllEqual(n float64) bool {
 
 	if (q.Len()<q.MaxItems()) {
 		return false;
